Update Game.java
I added the option to change the difficulty of the game under the settings menu. When doing so the user generates a new game. I was thinking of adding a dialog box letting the user know that when selecting a new difficulty it will erase any progress (unless saved) and prompt them to continue or cancel.