Update Game.java
Added the new game functionality and created a dialog box alerting the user that any unsaved progress will be lost. They must then select yes to continue or no to cancel. (I can change the text later to say continue/cancel. The methods are also very repetitive, and I can work on combining them so it's a lot cleaner.