Update StartPage.java
Changed the start page to have a custom listener for the load button. When pressed, the load button will trigger the loadListener and automatically create a new game, then call the load method on that game. This should fix the issue where nothing shows up when load is selected from the start page.