Combined Documentation
Here is everyone's individual reports and test cases combined into the final report. All the diagrams have been added as well. The only sections that currently do not have any topics are "System Architecture and System Design", "User Interface Design and Implementations", and "Configuration and Installation". I will complete the first two by tomorrow night hopefully, but didn't know what to put for the last one because I wasn't sure if we had decided on making an applet or not. Thank god this is almost over.